feat(warnMissingVars): refactor global env var check function
