+++ tools/mandatory-env-vars.json
@@ -1,0 +1,17 @@
+[
+  { "key": "ADMIN_PASSWORD", "msg": "Used to authenticate various admin tasks, please enter." },
+  { "key": "SENDGRID_API_KEY", "msg": "API key for Sendgrid, the emailer service, is missing." },
+  { "key": "ALGOLIA_APP_ID", "msg": "API key for Algolia is missing." },
+  { "key": "ALGOLIA_API_KEY", "msg": "API key for Algolia is missing." },
+  { "key": "ALGOLIA_SEARCH_KEY", "msg": "API key for Algolia is missing." },
+  { "key": "JWPLAYER_KEY", "msg": "API key for the video player is missing." },
+  { "key": "JWPLAYER_SECRET", "msg": "API key for the video player is missing." },
+  {
+    "key": "GOOGLE_APPLICATION_CREDENTIALS",
+    "msg": "You have missing service account keys. Please see: https://www.notion.so/cascade8/Preparing-your-Environment-e3c184db24d447c496c3241d4f16bc94"
+  },
+  {
+    "key": "FIREBASE_CI_SERVICE_ACCOUNT",
+    "msg": "You have missing service account keys. Please see: https://www.notion.so/cascade8/Preparing-your-Environment-e3c184db24d447c496c3241d4f16bc94"
+  }
+]
