+++ tools/scripts/lib.ts
