--- tools/scripts/mandatory-env-vars.json
@@ -1,12 +1,0 @@
-{
-  "required": [
-    "CASCADE8_ADMIN",
-    "ADMIN_PASSWORD",
-    "SENDGRID_API_KEY",
-    "ALGOLIA_APP_ID",
-    "ALGOLIA_API_KEY",
-    "ALGOLIA_SEARCH_KEY",
-    "JWPLAYER_KEY",
-    "JWPLAYER_SECRET"
-  ]
-}
